+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Fixed site links being present in a title of paid results
 - Better explanation about max limit of results
 - Added [Apify](https://sdk.apify.com/docs/api/apify) to `customDataFunction`
+- Added `includeUnfilteredResults` option